+++ js/data.js
@@ -1,0 +1,113 @@
+import { getRandomPositiveInteger, getRandomArrayElement } from'./util.js';
+
+// именаa
+const NAMES = [
+  'Ева',
+  'Александр',
+  'Глеб',
+  'Кира',
+  'Виктория',
+  'Анна',
+  'Оливия',
+  'Вероника',
+  'Михаил',
+  'Дарья',
+  'Матвей',
+  'Алиса',
+  'Александра',
+  'Кристина',
+  'Мария',
+  'Ксения',
+  'Роберт',
+  'Сафия',
+  'Софья',
+  'Дамир',
+  'Савва',
+  'Елена',
+  'Аврора',
+  'Варвара',
+  'Артём',
+];
+
+// comments
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо',
+  'Но не всё',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра',
+  'В конце концов это просто непрофессионально',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше',
+  'Лица у людей на фотке перекошены, как будто их избивают.',
+  'Как можно было поймать такой неудачный момент?!',
+];
+
+// описание под фото
+const DESCRIPTION_PHOTO = [
+  'Персонаж решил повыпендриваться перед своей возлюбленной, подкатив к ней на коленях со стаканом чего-нибудь прямо в разгар вечеринки!',
+  'Купили понравившейся девушке 101 розу, вложили кучу денег, а у неё аллергия на цветы...',
+  'Дневник персонажа прочитали',
+  'Хотел погладить животное, но оно укусило персонажа',
+  'Персонаж, случайно ударив кого-то, не смог определится что ему сказать: "тебе больно?" или "мне жаль!", после чего, запаниковав, выкрикнул " Тебе жаль??"',
+  'Кепку персонажа сдувает и уносит ветром.',
+  'Персонаж при падении утягивает другого персонажа за собой. В итоге произошёл поцелуй',
+  'Персонаж долго не ел и в итоге потерял сознание перед любимым(ой), с которым(ой) только удалось поговорить',
+  'Персонаж внезапно теряется в пространстве,не понимает куда он двигается',
+  'Персонаж во время речи случайно падает со сцены.',
+  'Персонаж называет свою девушку другим именем',
+  'Персонаж выбрасывает что-то важное.',
+  'Персонаж слишком много ест и набирает на банкете много еды в свою тарелку, чем вызывает недовольство окружающих.',
+];
+
+// количество фотографий
+const NUMBER_OF_PHOTOS = 25;
+
+// минимальная переменная в виде 0
+const minInt = 0;
+
+// количество разных аватаров.
+const avatars = 6;
+// макс. количество комментариев.
+const maxComments = 3;
+
+//условие мин. и макс. количество лайков.
+const minLikes = 15;
+const maxLikes = 200;
+
+// счетсик ID (const  -  выдвал  ошибку. Почему?)
+let idCounter = 0;
+
+// счетсик ID комментариев
+let idCommentsCounter = 0;
+
+// сущьность коментаторов.
+const generateComment = () => ({
+  id: ++idCounter,
+  avatar: `img/avatar-${getRandomPositiveInteger(minInt, avatars)}.svg.` ,
+  name: getRandomArrayElement(NAMES),
+  message: getRandomArrayElement(MESSAGES),
+});
+
+// сущьность фотографии.
+const generatePhoto = () => ({
+  id: ++idCommentsCounter,
+  url: `photos/${idCounter}.jpg`,
+  description: getRandomArrayElement(DESCRIPTION_PHOTO),
+  likes: getRandomPositiveInteger(minLikes, maxLikes),
+  comments: Array.from({ length: getRandomPositiveInteger(minInt, maxComments)}, generateComment),
+});
+
+const numberOfPhotos = Array.from({ length: NUMBER_OF_PHOTOS}, generatePhoto);
+
+// function showPhots() {
+//   return(numberOfPhotos);
+// }
+
+function showPhots() {
+  for (let i = 0; i < numberOfPhotos.length; i++) {
+    return(numberOfPhotos[i]);
+  }
+}
+
+showPhots(numberOfPhotos);
+export { showPhots };
