+++ js/main.js
@@ -1,105 +1,5 @@
-// имена
-const NAMEs = [
-  'Ева',
-  'Александр',
-  'Глеб',
-  'Кира',
-  'Виктория',
-  'Анна',
-  'Оливия',
-  'Вероника',
-  'Михаил',
-  'Дарья',
-  'Матвей',
-  'Алиса',
-  'Александра',
-  'Кристина',
-  'Мария',
-  'Ксения',
-  'Роберт',
-  'Сафия',
-  'Софья',
-  'Дамир',
-  'Савва',
-  'Елена',
-  'Аврора',
-  'Варвара',
-  'Артём',
-];
+import './data.js';
+import './util.js';
+import { showPhots } from './data.js';
 
-// comments
-const MESSAGEs = [
-  'Всё отлично!',
-  'В целом всё неплохо',
-  'Но не всё',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра',
-  'В конце концов это просто непрофессионально',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше',
-  'Лица у людей на фотке перекошены, как будто их избивают.',
-  'Как можно было поймать такой неудачный момент?!',
-];
-
-// описание под фото
-const DESCRIPTION_PHOTO = [
-  'Персонаж решил повыпендриваться перед своей возлюбленной, подкатив к ней на коленях со стаканом чего-нибудь прямо в разгар вечеринки!',
-  'Купили понравившейся девушке 101 розу, вложили кучу денег, а у неё аллергия на цветы...',
-  'Дневник персонажа прочитали',
-  'Хотел погладить животное, но оно укусило персонажа',
-  'Персонаж, случайно ударив кого-то, не смог определится что ему сказать: "тебе больно?" или "мне жаль!", после чего, запаниковав, выкрикнул " Тебе жаль??"',
-  'Кепку персонажа сдувает и уносит ветром.',
-  'Персонаж при падении утягивает другого персонажа за собой. В итоге произошёл поцелуй',
-  'Персонаж долго не ел и в итоге потерял сознание перед любимым(ой), с которым(ой) только удалось поговорить',
-  'Персонаж внезапно теряется в пространстве,не понимает куда он двигается',
-  'Персонаж во время речи случайно падает со сцены.',
-  'Персонаж называет свою девушку другим именем',
-  'Персонаж выбрасывает что-то важное.',
-  'Персонаж слишком много ест и набирает на банкете много еды в свою тарелку, чем вызывает недовольство окружающих.',
-];
-
-// колличество фотографий
-const NUMBER_OF_PHOTOS = 25;
-
-// функция целочисленных, позитивных, рандоманх чисел.
-function getRandomPositiveInteger (a, b) {
-  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
-  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
-  const result = Math.random() * (upper - lower + 1) + lower;
-
-  return Math.floor(result);
-}
-
-// счетсик ID (const  -  выдвал  ошибку. Почему?)
-let idCounter = 0;
-
-// счетсик ID комментариев
-let idCommentsCounter = 0;
-
-// функция которая возвращает случа эллемент массива на основе функции  которая возвращает случайное  число и  записывает в  объект.
-const getRandomArrayElement = (elements) => (elements[getRandomPositiveInteger(0, elements.length - 1)]);
-
-// сущьность коментаоров.
-const generateComment = () => ({
-  id: ++idCounter,
-  avatar: `img/avatar-${getRandomPositiveInteger(0, 6)}.svg.` ,
-  name: getRandomArrayElement(NAMEs),
-  message: getRandomArrayElement(MESSAGEs),
-});
-
-// сущьность фотографии.
-const generatePhoto = () => ({
-  id: ++idCommentsCounter,
-  url: `photos/${idCounter}.jpg`,
-  description: getRandomArrayElement(DESCRIPTION_PHOTO),
-  likes: getRandomPositiveInteger(15, 200),
-  comments: Array.from({ length: getRandomPositiveInteger(0, 3)}, generateComment),
-});
-
-const numberOfPhotos = Array.from({ length: NUMBER_OF_PHOTOS}, generatePhoto);
-
-function showArryPhots() {
-  return(numberOfPhotos);
-}
-showArryPhots();
-
-
+showPhots();
