+++ js/util.js
@@ -1,0 +1,13 @@
+// функция целочисленных, позитивных, рандоманх чисел.
+function getRandomPositiveInteger (a, b) {
+  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+}
+
+// функция которая возвращает случа эллемент массива на основе функции  которая возвращает случайное  число и  записывает в объект.
+const getRandomArrayElement = (elements) => (elements[getRandomPositiveInteger(0, elements.length - 1)]);
+
+export {getRandomPositiveInteger, getRandomArrayElement };
